Rewrite organize_scan_measurements
Previously, a temporary shots_dict was added to the BurstData object in
the bursts dict, but this isn't allowed because BurstData is a (immutable)
NamedTuple.

To remove the temporary shot_dict attribute (in order to keep BurstData
a NamedTuple), the Row objects are now organized in a (burst) dict of
(shot) dicts first, then the dict of dicts is converted to a list of
BurstData (which contains list of ShotData in its shots attribute).

Also adds MeasurementRow type hint.