Fix color scale on pointing and spectrum images, put in config
Make the color scales on the pointing and spectrum images, as well as the
y-axis on the spectrum lineout, constant across all images in the data
browser. This way intensities can be visually compared.

These settings:
* pointing_intensity_max
* spectrum_intensity_max
* spectrum_lineout_max

Can be saved and loaded in the UserData configuration, though this isn't
active yet.