Make measurement_db_engine a global object
Previously, sqlalchemy_engine and tables were module-level objects in
scans.py. Thus there was no way to update them if for example the
database settings changed.

This commit creates a MeasurementDBEngine class and an instance
measurement_db_engine in measurement_db.py, which holds the sqlalchemy
engine and tables and provides methods to update them.

It updates functions in scans.py to use this object.