Add preprocess function for spectrum images before 2024-11-01
image-processing-backend/rq-worker:1231fa7 was deployed on
2024-11-02 01:06:32 UTC, which includes (elementary) background subtraction
and normalization to energy and angle bin widths. Before that, only the
1D spectrum uses background subtraction and normalization.

This commit adds preprocessing for pre-2024-11-01 images so that those
show up properly in the data browser.

The background and scale values come from a fit done on 2024-11-05.

Adds utils/functools.py for a compose function.