Update background subtraction according to new fit
Re-ran fit program that calculates the image-wide background in the low-
and high-energy images as well as the scale factor to equalize the
brightness ("charge" calibration), this time for bursts between
2024-10-21T22-19-23 and 2024-10-30T00-17-57. One shot from each burst is
included.

The latter was previously something close to 2024-11-02T01:06:32Z, the
time when the image backend was updated to output background-subtracted
and energy-bin-width-normalized low and high energy spectrum images.

It is now set to the last time before magnitude of the low and high
energy images increased by 1000 for unknown reasons, before coming back
in bursts starting 2024-10-30T22-24-26.

The former is after major updates to the eSpecAnalysis algorithm.