Plot all of low energy image and only non-overlapping part of high
It turns out that the low and high energy images don't stitch together
well along the angle axis. We concluded that this is because of scattering
by the DRZ screen. So at this point, don't bother stitching.

The angle-integrated spectrum is still merged together, without
additional equalization beyond the global intensity scale factor.